High score now saves to external file
All in one commit! The game will create a folder in AppData if doesn't exist, and will save the High Score any time the player dies. Currently, only High Score is saved to file, current score is tied to PlayerPrefs but that can be changed if needed. I also renamed "CollectionManager" to "ScoreManager" to make it more clear, but this means some other files are affected.
